Pick-up Bug Fix, and Reworked Camera Controls
Pick-up bug:
When changing to Item hold states, it still tries to change the current object being held because it is programmed to get the pickup object every time the state starts.

Fix:
Made some conditions to only change the object when these conditons are met:
When the pick up object is null,
When the object nearest is NOT the pick up object,
Lastly, only when the current state is NOT Item hold.