ADDED Joystick for Mobile Movement, Fixed Jittery Camera
ADDED "Joystick Pack" by Fenerax Studios, from the Unity asset store.

Jittery Camera Cause: When setting the Y position of the player to the Y position of the Hit point, the Y Velocity is not set to 0 causing the player to fall a little causing the Y position of the player to change, and because of that, the Y position of the player must be set to the Y position of the hit point AGAIN. That cycle causes the jittery camera.

Fix: Set the Y velocity of the player to 0 before setting the Y position of the player to the Y position of the hit point.